Purchase cars functionality added (#10)
* Refactored Car subclasses constructors by removing long parameter lists.

* JavaDoc to Hatchback,Pickup,SUV,Sedan constructors.

* Filter cars functionality added

* Consistent Car Printing

* mend

* Purchase car implemented and updates CSV appropriately.

* Changed return type of getCarsPurchased() from double to int.

* Update number of cars purchased by user utilizing setter and getter.

* Only update count of vehicles to 0 in CSV. Do not delete contents of cSV.

+++ src/entity/User.java
@@ -32,7 +32,7 @@ public class User extends Person {
         this.balance = balance;
     }
 
-    public double getCarsPurchased() {
+    public int getCarsPurchased() {
         return carsPurchased;
     }
